Add alt mappings; Reorganize the file; Add plugins
Map alt [] to tmux's copy and paste. Map alt 5' to split % ". Map alt S to
synchronize panes. Map alt H to force split on the right. Add tmux
ressurect and tmux continuum via TPM (Tmux Plugin Manager), which is
also added in this commit. Add notes to the end of the file.